오류 수정
PC-노트북 사이에서 PC에서는 오류가 발생하지 않았으나 노트북에서만 발생하던 오류를 수정함

+++ Server/Server/FileMangement.cpp
@@ -27,7 +27,7 @@ void SaveFile(vector<Account>& v) {
 
 	// 먼저 컨테이너의 크기를 txt로 저장한다.
 	ofstream outSize("AcccountNum.txt");
-	unsigned int fsize{ v.size() };
+	unsigned int fsize{ (unsigned int)v.size() };
 	outSize << fsize;
 
 	// 컨테이너의 내용을 저장
